[Refactor] isolate common codes from watchers
