+++ src/base/common/file/file.ts
@@ -2,7 +2,7 @@ import { IDisposable } from "src/base/common/dispose";
 import { Register } from "src/base/common/event";
 import { IReadableStreamEvent } from "src/base/common/file/stream";
 import { URI } from "src/base/common/file/uri";
-import { IRawResourceChangeEvents } from "src/code/platform/files/node/watcher";
+import { IRawResourceChangeEvents } from "src/code/platform/files/common/watcher";
 
 export const enum ByteSize {
 	KB = 1024,
