+++ src/code/platform/files/common/resourceChangeEvent.ts
@@ -1,9 +1,8 @@
 import { URI } from "src/base/common/file/uri";
 import { CreateTernarySearchTree, TernarySearchTree } from "src/base/common/util/ternarySearchTree";
-import { IRawResourceChangeEvents, ResourceChangeType } from "src/code/platform/files/node/watcher";
+import { IRawResourceChangeEvent, IRawResourceChangeEvents, ResourceChangeType } from "src/code/platform/files/common/watcher";
 import { IReviverRegistrant } from "src/code/platform/ipc/common/revive";
 import { REGISTRANTS } from "src/code/platform/registrant/common/registrant";
-import { IRawResourceChangeEvent } from "src/code/platform/files/node/watcher";
 
 /**
  * An interface only for {@link ResourceChangeEvent}.
